WIP: Adding exception cases for key handling (F1~F12, Shift, Ctrl, etc)
