+++ examples/basic/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0, minimal-ui, viewport-fit=cover">
     <script src="https://aframe.io/releases/0.9.0/aframe.min.js"></script>
-    <script>window.AFK || document.write('<script src="../aframe-keyboard.min.js">\x3C/script>')</script>
+    <script>window.AFK || document.write('<script src="../../dist/aframe-keyboard.min.js">\x3C/script>')</script>
     <title>Keyboard Sample</title>
     <style>
       body {
