+++ src/components/aframe-keyboard.js
@@ -1,4 +1,6 @@
 /* global AFRAME, AFK */
+const validateKeyboardInput = require('../utils/events');
+
 if (typeof AFRAME === 'undefined') {
   throw new Error('Component attempted to register before AFRAME was available.');
 }
@@ -31,80 +33,11 @@ AFRAME.registerComponent('a-keyboard', {
   },
 
   handleKeyboardPress: function(e) {
-    let inputEvent;
-    const code = e.key.charCodeAt(0);
-    switch (e.keyCode) {
-      case 9: // tab
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code: e.keyCode, value: '\t'}});
-        document.dispatchEvent(inputEvent);
-        break;
-      case 8: // backspace
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code: e.keyCode, value: ''}});
-        document.dispatchEvent(inputEvent);
-        break;
-      case 13: // enter
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code: e.keyCode, value: '\n'}});
-        document.dispatchEvent(inputEvent);
-        break;
-      case 16: // shift
-        break;
-      case 18: // alt
-        break;
-      case 27: // esc
-        if (this.dismissable) {
-          inputEvent = new CustomEvent('a-keyboard-update', {detail: {code: e.keyCode, value: ''}});
-          document.dispatchEvent(inputEvent);
-        }
-        break;
-      case 32: // space
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code: e.keyCode, value: ' '}});
-        document.dispatchEvent(inputEvent);
-        break;
-      default:
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code, value: e.key}});
-        document.dispatchEvent(inputEvent);
-        break;
-    }
-
-    // Artificially trigger keypress events
-    const key = document.querySelector(`#a-keyboard-${code}`) || document.querySelector(`#a-keyboard-${e.keyCode}`);
-    if (key) {
-      key.dispatchEvent(new Event('mousedown'));
-      setTimeout(function() {
-        key.dispatchEvent(new Event('mouseleave'));
-      }, 80);
-    }
+    validateKeyboardInput(e);
   },
 
   handleKeyboardVR: function(e) {
-    let inputEvent;
-    const value = document.querySelector(`#${e.target.id}`).getAttribute('value');
-    const code = document.querySelector(`#${e.target.id}`).getAttribute('key-code');
-    switch (code) {
-      case '13': // enter
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code, value: '\n'}});
-        document.dispatchEvent(inputEvent);
-        break;
-      case '8': // backspace
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code: 8, value: ''}});
-        document.dispatchEvent(inputEvent);
-        break;
-      case '9': // tab
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code, value: '\t'}});
-        document.dispatchEvent(inputEvent);
-        break;
-      case '16': // shift
-        AFK.template.toggleActiveMode('shift');
-        break;
-      case '32': // space
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code, value: ' '}});
-        document.dispatchEvent(inputEvent);
-        break;
-      default:
-        inputEvent = new CustomEvent('a-keyboard-update', {detail: {code, value}});
-        document.dispatchEvent(inputEvent);
-        break;
-    }
+    validateKeyboardInput(e, 'vr');
   },
 
   remove: function() {
