--- src/i18n/kr.js
@@ -1,181 +1,0 @@
-module.exports = {
-  name: 'ms-US-International',
-  normal: [
-    [
-      {value: '1', type: 'standard'},
-      {value: '2', type: 'standard'},
-      {value: '3', type: 'standard'},
-      {value: '4', type: 'standard'},
-      {value: '5', type: 'standard'},
-      {value: '6', type: 'standard'},
-      {value: '7', type: 'standard'},
-      {value: '8', type: 'standard'},
-      {value: '9', type: 'standard'},
-      {value: '0', type: 'standard'},
-      {value: '<-', type: 'delete'},
-    ],
-    [
-      {value: 'ㅂ', type: 'standard'},
-      {value: 'ㅈ', type: 'standard'},
-      {value: 'ㄷ', type: 'standard'},
-      {value: 'ㄱ', type: 'standard'},
-      {value: 'ㅅ', type: 'standard'},
-      {value: 'ㅛ', type: 'standard'},
-      {value: 'ㅕ', type: 'standard'},
-      {value: 'ㅑ', type: 'standard'},
-      {value: 'ㅐ', type: 'standard'},
-      {value: 'ㅔ', type: 'standard'},
-      {value: 'Ent', type: 'enter'},
-    ],
-    [
-      {value: 'ㅁ', type: 'standard'},
-      {value: 'ㄴ', type: 'standard'},
-      {value: 'ㅇ', type: 'standard'},
-      {value: 'ㄹ', type: 'standard'},
-      {value: 'ㅎ', type: 'standard'},
-      {value: 'ㅗ', type: 'standard'},
-      {value: 'ㅓ', type: 'standard'},
-      {value: 'ㅏ', type: 'standard'},
-      {value: 'ㅣ', type: 'standard'},
-      {value: '"', type: 'standard'},
-      // last one empty because Enter key spans 2 rows
-    ],
-    [
-      {value: 'Shift', type: 'shift'},
-      {value: 'ㅋ', type: 'standard'},
-      {value: 'ㅌ', type: 'standard'},
-      {value: 'ㅊ', type: 'standard'},
-      {value: 'ㅍ', type: 'standard'},
-      {value: 'ㅠ', type: 'standard'},
-      {value: 'ㅜ', type: 'standard'},
-      {value: 'ㅡ', type: 'standard'},
-      {value: 'Alt', type: 'alt'},
-    ],
-    [
-      {value: 'Cancel', type: 'cancel'},
-      {value: '\u2423', type: 'space'},
-      {value: ',', type: 'standard'},
-      {value: '.', type: 'standard'},
-      {value: 'Submit', type: 'submit'},
-    ],
-  ],
-  shift: [
-    [
-      {value: '1', type: 'standard'},
-      {value: '2', type: 'standard'},
-      {value: '3', type: 'standard'},
-      {value: '4', type: 'standard'},
-      {value: '5', type: 'standard'},
-      {value: '6', type: 'standard'},
-      {value: '7', type: 'standard'},
-      {value: '8', type: 'standard'},
-      {value: '9', type: 'standard'},
-      {value: '0', type: 'standard'},
-      {value: '\u232B', type: 'delete'},
-    ],
-    [
-      {value: 'ㅃ', type: 'standard'},
-      {value: 'ㅉ', type: 'standard'},
-      {value: 'ㄸ', type: 'standard'},
-      {value: 'ㄲ', type: 'standard'},
-      {value: 'ㅆ', type: 'standard'},
-      {value: 'ㅛ', type: 'standard'},
-      {value: 'ㅕ', type: 'standard'},
-      {value: 'ㅑ', type: 'standard'},
-      {value: 'ㅒ', type: 'standard'},
-      {value: 'ㅖ', type: 'standard'},
-      {value: '\u23CE', type: 'enter'},
-    ],
-    [
-      {value: 'ㅁ', type: 'standard'},
-      {value: 'ㄴ', type: 'standard'},
-      {value: 'ㅇ', type: 'standard'},
-      {value: 'ㄹ', type: 'standard'},
-      {value: 'ㅎ', type: 'standard'},
-      {value: 'ㅗ', type: 'standard'},
-      {value: 'ㅓ', type: 'standard'},
-      {value: 'ㅏ', type: 'standard'},
-      {value: 'ㅣ', type: 'standard'},
-      {value: '"', type: 'standard'},
-      // last one empty because Enter key spans 2 rows
-    ],
-    [
-      {value: '\u21E7', type: 'shift'},
-      {value: 'ㅋ', type: 'standard'},
-      {value: 'ㅌ', type: 'standard'},
-      {value: 'ㅊ', type: 'standard'},
-      {value: 'ㅍ', type: 'standard'},
-      {value: 'ㅠ', type: 'standard'},
-      {value: 'ㅜ', type: 'standard'},
-      {value: 'ㅡ', type: 'standard'},
-      {value: 'Alt', type: 'alt'},
-    ],
-    [
-      {value: 'Cancel', type: 'cancel'},
-      {value: '\u2423', type: 'space'},
-      {value: ',', type: 'standard'},
-      {value: '.', type: 'standard'},
-      {value: 'Submit', type: 'submit'},
-    ],
-  ],
-  alt: [
-    [
-      {value: '1', type: 'standard'},
-      {value: '2', type: 'standard'},
-      {value: '3', type: 'standard'},
-      {value: '4', type: 'standard'},
-      {value: '5', type: 'standard'},
-      {value: '6', type: 'standard'},
-      {value: '7', type: 'standard'},
-      {value: '8', type: 'standard'},
-      {value: '9', type: 'standard'},
-      {value: '0', type: 'standard'},
-      {value: '\u232B', type: 'delete'},
-    ],
-    [
-      {value: '~', type: 'standard'},
-      {value: '`', type: 'standard'},
-      {value: '|', type: 'standard'},
-      {value: '•', type: 'standard'},
-      {value: '√', type: 'standard'},
-      {value: 'π', type: 'standard'},
-      {value: 'u', type: 'standard'},
-      {value: '÷', type: 'standard'},
-      {value: 'x', type: 'standard'},
-      {value: '-', type: 'standard'},
-      {value: '\u23CE', type: 'enter'},
-    ],
-    [
-      {value: '₤', type: 'standard'},
-      {value: '#', type: 'standard'},
-      {value: '€', type: 'standard'},
-      {value: '°', type: 'standard'},
-      {value: '^', type: 'standard'},
-      {value: '_', type: 'standard'},
-      {value: '=', type: 'standard'},
-      {value: '[', type: 'standard'},
-      {value: ']', type: 'standard'},
-      {value: '!', type: 'standard'},
-      // last one empty because Enter key spans 2 rows
-    ],
-    [
-      {value: '\u21E7', type: 'shift'},
-      {value: '@', type: 'standard'},
-      {value: '#', type: 'standard'},
-      {value: '$', type: 'standard'},
-      {value: '%', type: 'standard'},
-      {value: '*', type: 'standard'},
-      {value: '(', type: 'standard'},
-      {value: ')', type: 'standard'},
-      {value: '-', type: 'standard'},
-      {value: '?', type: 'standard'},
-    ],
-    [
-      {value: 'Cancel', type: 'cancel'},
-      {value: 'Alt', type: 'alt'},
-      {value: '\u2423', type: 'space'},
-      {value: '@', type: 'standard'},
-      {value: 'Submit', type: 'submit'},
-    ],
-  ],
-};
